code_quality_analyzer_analysis/requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
exceptiongroup==1.1.3
iniconfig==2.0.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pluggy==1.3.0
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
